Button System Update
added button backs which allow for red buttons. I also made it so icons are placed in the loop queue, the loop queue will have to be changed  to include less symbols because 10 seems too big for a phone screen. A smaller change is that individual objects can be set to have different overlap thresholds